Capitalized classes file names
