+++ fe1-web/Model/Network/Method/Broadcast.ts
@@ -1,0 +1,3 @@
+import { JsonRpcParamsWithMessage } from './JsonRpcParams';
+
+export class Broadcast extends JsonRpcParamsWithMessage {}
