+++ fe1-web/Model/Network/Method/Catchup.ts
@@ -1,0 +1,3 @@
+import { JsonRpcParams } from './JsonRpcParams';
+
+export class Catchup extends JsonRpcParams {}
