+++ fe1-web/Model/Network/Method/Message/data/rollCall/CloseRollCall.ts
