+++ fe1-web/Model/Network/Method/Message/data/rollCall/index.ts
@@ -1,3 +1,3 @@
-export * from './createRollCall';
-export * from './closeRollCall';
-export * from './openRollCall';
+export * from './CreateRollCall';
+export * from './CloseRollCall';
+export * from './OpenRollCall';
