+++ fe1-web/Model/Network/Method/Message/data/witness/index.ts
@@ -1,1 +1,1 @@
-export * from './witnessMessage';
+export * from './WitnessMessage';
