+++ fe1-web/Model/Network/Method/Message/index.ts
@@ -1,1 +1,1 @@
-export * from './message';
+export * from './Message';
