+++ fe1-web/Model/Network/Method/Publish.ts
@@ -1,0 +1,3 @@
+import { JsonRpcParamsWithMessage } from './JsonRpcParams';
+
+export class Publish extends JsonRpcParamsWithMessage {}
