+++ fe1-web/Model/Network/Method/Subscribe.ts
@@ -1,0 +1,3 @@
+import { JsonRpcParams } from './JsonRpcParams';
+
+export class Subscribe extends JsonRpcParams {}
