+++ fe1-web/Model/Network/Method/Unsubscribe.ts
@@ -1,0 +1,3 @@
+import { JsonRpcParams } from './JsonRpcParams';
+
+export class Unsubscribe extends JsonRpcParams {}
