--- fe1-web/Model/Network/Method/broadcast.ts
@@ -1,3 +1,0 @@
-import { JsonRpcParamsWithMessage } from './jsonRpcParams';
-
-export class Broadcast extends JsonRpcParamsWithMessage {}
