--- fe1-web/Model/Network/Method/catchup.ts
@@ -1,3 +1,0 @@
-import { JsonRpcParams } from './jsonRpcParams';
-
-export class Catchup extends JsonRpcParams {}
