+++ fe1-web/Model/Network/Method/index.ts
@@ -1,5 +1,5 @@
-export * from './broadcast';
-export * from './catchup';
-export * from './publish';
-export * from './subscribe';
-export * from './unsubscribe';
+export * from './Broadcast';
+export * from './Catchup';
+export * from './Publish';
+export * from './Subscribe';
+export * from './Unsubscribe';
