--- fe1-web/Model/Network/Method/publish.ts
@@ -1,3 +1,0 @@
-import { JsonRpcParamsWithMessage } from './jsonRpcParams';
-
-export class Publish extends JsonRpcParamsWithMessage {}
