--- fe1-web/Model/Network/Method/subscribe.ts
@@ -1,3 +1,0 @@
-import { JsonRpcParams } from './jsonRpcParams';
-
-export class Subscribe extends JsonRpcParams {}
