--- fe1-web/Model/Network/Method/unsubscribe.ts
@@ -1,3 +1,0 @@
-import { JsonRpcParams } from './jsonRpcParams';
-
-export class Unsubscribe extends JsonRpcParams {}
