+++ fe1-web/Model/Network/index.ts
@@ -1,4 +1,4 @@
 export * from './jsonRpcMethods';
-export * from './jsonRpcRequest';
-export * from './jsonRpcResponse';
+export * from './JsonRpcRequest';
+export * from './JsonRpcResponse';
 export * from './ProtocolError';
